+++ src/Rules/Different.php
@@ -15,12 +15,19 @@ class Different extends Rule
     protected string $message = 'rule.different';
     protected array $fillableParams = ['field'];
 
+    public function field(string $field): self
+    {
+        $this->params['field'] = $field;
+
+        return $this;
+    }
+
     public function check($value): bool
     {
         $this->assertHasRequiredParameters($this->fillableParams);
 
         $field        = $this->parameter('field');
-        $anotherValue = $this->validation->getValue($field);
+        $anotherValue = $this->validation->input()->get($field);
 
         return $value != $anotherValue;
     }