+++ src/Rules/ProhibitedIf.php
@@ -40,7 +40,7 @@ class ProhibitedIf extends Rule
         $definedValues    = $this->parameter('values');
         $anotherValue     = $this->attribute()->value($anotherAttribute);
 
-        $requiredValidator = $this->validation->getFactory()->rule('required');
+        $requiredValidator = $this->validation->factory()->rule('required');
 
         if (in_array($anotherValue, $definedValues, is_bool($anotherValue))) {
             return !$requiredValidator->check($value);