+++ src/Rules/ProhibitedUnless.php
@@ -41,7 +41,7 @@ class ProhibitedUnless extends Rule
         $definedValues    = $this->parameter('values');
         $anotherValue     = $this->attribute()->value($anotherAttribute);
 
-        $requiredValidator = $this->validation->getFactory()->rule('required');
+        $requiredValidator = $this->validation->factory()->rule('required');
 
         if (!in_array($anotherValue, $definedValues, is_bool($anotherValue))) {
             return !$requiredValidator->check($value);