+++ src/Rules/Required.php
@@ -23,7 +23,7 @@ class Required extends Rule
         $this->setAttributeAsRequired();
 
         if ($this->attribute?->hasRule('uploaded_file')) {
-            return $this->isValueFromUploadedFiles($value) and $value['error'] != UPLOAD_ERR_NO_FILE;
+            return $this->isValueFromUploadedFiles($value) && $value['error'] != UPLOAD_ERR_NO_FILE;
         }
 
         if (is_string($value)) {