+++ src/Rules/Same.php
@@ -15,6 +15,13 @@ class Same extends Rule
     protected string $message = 'rule.same';
     protected array $fillableParams = ['field'];
 
+    public function field(string $field): self
+    {
+        $this->params['field'] = $field;
+
+        return $this;
+    }
+
     public function check($value): bool
     {
         $this->assertHasRequiredParameters($this->fillableParams);